+++ gldas/interface.py
@@ -31,14 +31,13 @@ except ImportError:
 
 from pygeobase.io_base import ImageBase, MultiTemporalImageBase
 from pygeobase.object_base import Image
-import pygeogrids
 from pynetcf.time_series import GriddedNcOrthoMultiTs
 
 from datetime import timedelta
 
 from gldas.grid import GLDAS025Cellgrid
 from netCDF4 import Dataset
-
+from pygeogrids.netcdf import load_grid
 
 
 class GLDAS_Noah_v21_025Img(ImageBase):
@@ -414,7 +413,7 @@ class GLDASTs(GriddedNcOrthoMultiTs):
         if grid_path is None:
             grid_path = os.path.join(ts_path, "grid.nc")
 
-        grid = pygeogrids.netcdf.load_grid(grid_path)
+        grid = load_grid(grid_path)
         super(GLDASTs, self).__init__(ts_path, grid)