+++ tests/test_download.py
@@ -12,47 +12,47 @@ from gldas.download import gldas_folder_get_version_first_last
 def test_get_last_dir_in_dir():
     path = os.path.join(os.path.dirname(__file__),
                         'folder_test', 'success')
-    last_dir = get_last_formatted_dir_in_dir(path, "{:%Y}")
+    last_dir = get_last_formatted_dir_in_dir(path, "{time:%Y}")
     assert last_dir == '2014'
 
 def test_get_last_dir_in_dir_failure():
     path = os.path.join(os.path.dirname(__file__),
                         'folder_test', 'failure')
-    last_dir = get_last_formatted_dir_in_dir(path, "{:%Y}")
+    last_dir = get_last_formatted_dir_in_dir(path, "{time:%Y}")
     assert last_dir == None
 
 
 def test_get_first_dir_in_dir():
     path = os.path.join(os.path.dirname(__file__),
                         'folder_test', 'success')
-    last_dir = get_first_formatted_dir_in_dir(path, "{:%Y}")
+    last_dir = get_first_formatted_dir_in_dir(path, "{time:%Y}")
     assert last_dir == '2013'
 
 def test_get_last_gldas_folder():
     path = os.path.join(os.path.dirname(__file__),
                         'folder_test', 'success')
-    last = get_last_gldas_folder(path, ['{:%Y}', '{:%j}'])
+    last = get_last_gldas_folder(path, ['{time:%Y}', '{time:%j}'])
     last_should = os.path.join(path, "2014", "134")
     assert last == last_should
 
 def test_get_last_gldas_folder_no_folder():
     path = os.path.join(os.path.dirname(__file__),
                         'folder_test', 'failure')
-    last = get_last_gldas_folder(path, ['{:%Y}', '{:%j}'])
+    last = get_last_gldas_folder(path, ['{time:%Y}', '{time:%j}'])
     last_should = None
     assert last == last_should
 
 def test_get_first_gldas_folder():
     path = os.path.join(os.path.dirname(__file__),
                         'folder_test', 'success')
-    last = get_first_gldas_folder(path, ['{:%Y}', '{:%j}'])
+    last = get_first_gldas_folder(path, ['{time:%Y}', '{time:%j}'])
     last_should = os.path.join(path, "2013", "001")
     assert last == last_should
 
 def test_get_first_gldas_folder_no_folder():
     path = os.path.join(os.path.dirname(__file__),
                         'folder_test', 'failure')
-    last = get_first_gldas_folder(path, ['{:%Y}', '{:%j}'])
+    last = get_first_gldas_folder(path, ['{time:%Y}', '{time:%j}'])
     last_should = None
     assert last == last_should
 
@@ -67,3 +67,15 @@ def test_gldas_get_start_end():
     assert end == end_should
     assert start == start_should
     
+
+
+if __name__ == '__main__':
+    test_gldas_get_start_end()
+    test_get_last_dir_in_dir()
+    test_get_first_dir_in_dir()
+    test_get_first_gldas_folder()
+    test_get_first_gldas_folder_no_folder()
+    test_get_last_dir_in_dir()
+    test_get_last_gldas_folder()
+    test_get_last_gldas_folder_no_folder()
+    test_get_last_dir_in_dir_failure()