+++ tests/test_interface.py
@@ -248,3 +248,7 @@ def test_GLDAS_Noah_v21_025Img_img_reading_2D():
 
 
 
+if __name__ == '__main__':
+    test_GLDAS_Noah_v1_025Ds_img_reading()
+
+